backend/requirements.txt
Flask
flask-cors
torch
Pillow
torchvision
python-dotenv
werkzeug
groq
